Fancy window bar buttons, added drag button, improved event handler
Removed obsolete drag window behavior for not allowing dragging when the source is a button or textbox. This is obsolete because the mouse click is being swallowed by the button or textbox so it doesn't trigger the eventhandler.

Added a label/border element for dragging the window. This is more intuitive for users who are mostly used to dragging a screen by holding left mouse button on the window bar.

Discovered some fancy gradual opacity increase animation. Applied it to the menu buttons because it's hella fancy!