Set starting volume to 50% to prevent too loud music.
Fixed the Previous/Next song behavior by making it aware of whether the player has played music before. To prevent the player from starting to play a song when user pressed the previous/next song button while the player is not playing we need to know if the player is playing. Since loading a new song requires the player to be stopped we can't simply listen to stopped. This player implementation will only pause and resume. So when we are paused or when we haven't started playing a song yet we shouldn't start playing a song when skipping to the previous/next song.

The volume slider thumb now only slides when dragging the thumb. The mouse moving behavior caused jitter because it interacted in conjunction with the thumb moving. I'm now relying on WPF basic behavior. It's not as fancy, but it's not janky and it reduces code complexity.

Dispose of disposables in both the music player viewmodel and the jukebox to manage memory correctly.