Improved Eventhub observable names and config is now stored in a file.
Improved names of the observables in the Eventhub. I didnt suffix them with Observable because it can be determined from the signature when writing new code and on existing code the subscription makes it clear that its an observable.

Configuration is now stored in a json file next to the .exe file. Upon loading the application the configuration is loaded from the json and applied.
When the user changes the configuration from the configuration sub dialog the config is applied and stored only when the config is stored.