Improved generic background color and GeneralConfig
Made the background of the innermost border the generic background color so that child elements can apply margin without messing up the background color.

Renamed the GeneralConfig service to a manager to make it clear what it does while also freeing up the model name. Changed it to allow for general configuration setting. This opened up the possibility to set multiple types of configuration in an easy way.