Food topic is wrapped up for now.
Moved more generic stuff to the base viewmodel for topics. Made the home screen a normal screen, it's not a topic and as such, it is lacking the default topic stuff.
The new topic base class should make implementing topics much easier. I don't want to make a template because at this point in time, I believe that the topic screens themselves are not generic across the board. I do not want to implement subtypes of topic screens just yet.